[Templates PR] - combine AppSetting & principal ID
