+++ AzureFunctions.AngularClient/src/app/app.module.ts
@@ -106,7 +106,6 @@ import { SiteConfigComponent } from './site/site-config/site-config.component';
 import { CommandBarComponent } from './controls/command-bar/command-bar.component';
 import { CommandComponent } from './controls/command-bar/command/command.component';
 import { EventHubComponent } from './pickers/event-hub/event-hub.component';
-import { MicrosoftGraphComponent } from './pickers/microsoft-graph/microsoft-graph.component';
 import { ServiceBusComponent } from './pickers/service-bus/service-bus.component';
 import { SlotsListComponent } from './slots-list/slots-list.component';
 import { SlotsService } from './shared/services/slots.service';
@@ -215,7 +214,6 @@ export class AppModule {
       SlotsListComponent,
       SlotNewComponent,
       EventHubComponent,
-      MicrosoftGraphComponent,
       ServiceBusComponent,
       SearchBoxComponent,
       AppSettingComponent,
