+++ AzureFunctions.AngularClient/src/app/binding-input/binding-input.component.html
@@ -78,13 +78,18 @@
   </div>
 
   <div *ngIf="input?.type === 'picker'" [ngClass]="input?.class" class="picker-container">
-    <select id="{{input.id}}"
-            (ngModelChange)="inputChanged($event)"
-            [(ngModel)]="input.value"
-            [class.input-error]="!!input.errorText"
-            [class.hide-arrow]="!!input.errorText">
-      <option *ngFor="let item of input.items" [value]="item">{{item}}</option>
-    </select>
+    <!-- Show drop-down menu of potential app settings -->
+    <input type="text"
+          [(ngModel)]="input.value"
+          [attr.list]=input.id
+          (ngModelChange)="inputChanged($event)"
+          [class.input-error]="!!input.errorText"
+          [class.hide-arrow]="!!input.errorText"
+          [disabled]=input.isDisabled />
+    <!-- select + textbox -->
+    <datalist id="{{input.id}}">
+      <option *ngFor="let item of input.items" [value]="item">
+    </datalist>
     <pop-over *ngIf="!!input.errorText" [message]="input.errorText" [isInputError]="true">
     </pop-over>
     <!--[leftOffset]=20 [topOffset]=10-->
@@ -119,11 +124,4 @@
       <input [(ngModel)]="input.value[item.value]" (ngModelChange)="inputChanged($event)" type="checkbox" [fnWriteAccess]="functionApp"> {{ item.display }}
     </div>
   </div>
-
-  <microsoft-graph *ngIf="input?.type === 'appSetting'"
-                   [fnWriteAccess]="functionApp"
-                   [functionApp]="functionApp"
-                   [input]="input">
-  </microsoft-graph>
-
 </div>
