--- AzureFunctions.AngularClient/src/app/pickers/microsoft-graph/microsoft-graph.component.html
@@ -1,16 +1,0 @@
-﻿<div *ngIf="input?.type === 'appSetting'" class="picker-container">
-  <!-- Show drop-down menu of potential app settings -->
-  <input type="text"
-         [(ngModel)]="input.value"
-         [attr.list]=input.id
-         (ngModelChange)="inputChanged($event)"
-         [disabled]=input.isDisabled />
-
-  <datalist id="{{input.id}}">
-    <option *ngFor="let item of input.items" [value]="item">
-  </datalist>
-
-  <!-- Show link to retrieve principal id -->
-  <i *ngIf="!input.inProcess && !showTryView" class="link" [class.select]="input.errorText" (click)="openLogin(input)"> {{ 'bindingInput_retrieve' | translate }}</i>
-
-</div>
