--- AzureFunctions.AngularClient/src/app/pickers/microsoft-graph/microsoft-graph.component.spec.ts
@@ -1,24 +1,0 @@
-﻿import { AppModule } from './../../app.module';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { MicrosoftGraphComponent } from './microsoft-graph.component';
-
-describe('MicrosoftGraphComponent', () => {
-  let component: MicrosoftGraphComponent;
-  let fixture: ComponentFixture<MicrosoftGraphComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule(AppModule.moduleDefinition)
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(MicrosoftGraphComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
